UPSTREAM: media: rc: no need to announce major number
Since commit a60d64b15c20 ("media: lirc: lirc interface should not be
a raw decoder"), the message in the documentation is incorrect as the
module name is rc_core, not lirc_dev. Since the message is not useful,
just make the message debug and remove it from the documentation.

Change-Id: I945a6a2f99e27e8a62e62a3e8f5547b73de93c0a
Signed-off-by: Sean Young <sean@mess.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
Signed-off-by: Vayruz Rafli <mikhailovnakowalski@gmail.com>